mafia_a2a_game/requirements.txt
a2a-sdk
dotenv
uvicorn
click
httpx
asyncio
nest_asyncio
pydantic